+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 - Add ability to download attachments for comments
 - Increase default http timeout to 120s
 - Add `--resume-on-error` flag when creating annotations
+- Remove `--use-moon-forms` flag
 
 # v0.28.0
 - Add general fields to `create datasets`