+++ CHANGELOG.md
@@ -1,11 +1,10 @@
-# Unreleased
+# v0.29.0
 - Add `config parse-from-url` command for parsing configuration from a URL
 - Add ability to download attachments for comments
 - Increase default http timeout to 120s
 - Add `--resume-on-error` flag when creating annotations
 - Remove `--use-moon-forms` flag
 - Add `--resume-on-error` flag when creating comments / emails
-- Serialize assigned on moon_forms if empty vec
 
 # v0.28.0
 - Add general fields to `create datasets`